refactor: EntityInspectorの分割および細かな修正
