+++ packages/frontend/src/pages/world-editor/components/inspector/EntityInspector.tsx
@@ -1,8 +1,10 @@
-import type { EntityComponentDef, InitialEntity } from '@ubichill/shared';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import type { InitialEntity } from '@ubichill/shared';
 import { css } from '@/styled-system/css';
-import type { AvailableEntityKind, DataFieldSpec, DataFields } from '../../hooks/useAvailableEntityKinds';
-import { COMPONENT_DRAG_MIME } from '../../lib/dnd';
+import type { AvailableEntityKind } from '../../hooks/useAvailableEntityKinds';
+import { ComponentCard, ComponentPicker } from './ComponentCard';
+import { EntityHeader } from './EntityHeader';
+import { NumField, Section } from './primitives';
+import { TagsEditor } from './TagsEditor';
 
 interface EntityInspectorProps {
     entity: InitialEntity;
@@ -20,6 +22,10 @@ interface EntityInspectorProps {
     onRenameEntity: (newId: string) => void;
 }
 
+/**
+ * Entity 編集ペインのトップレベル composer。
+ * 各セクションは個別ファイルに分割済 (EntityHeader / TagsEditor / ComponentCard 等)。
+ */
 export function EntityInspector({
     entity,
     initiallyExpandedComponentIndex,
@@ -107,1086 +113,3 @@ export function EntityInspector({
         </div>
     );
 }
-
-// ============================================
-// Component カード (Unity 風アコーディオン)
-// ============================================
-
-interface ComponentCardProps {
-    component: EntityComponentDef;
-    componentIndex: number;
-    dataFields?: DataFields;
-    initiallyExpanded: boolean;
-    onChange: (updater: (prev: InitialEntity) => InitialEntity) => void;
-    onDelete: () => void;
-}
-
-function ComponentCard({
-    component,
-    componentIndex,
-    dataFields,
-    initiallyExpanded,
-    onChange,
-    onDelete,
-}: ComponentCardProps) {
-    const [expanded, setExpanded] = useState(initiallyExpanded);
-
-    useEffect(() => {
-        if (initiallyExpanded) setExpanded(true);
-    }, [initiallyExpanded]);
-
-    return (
-        <div
-            className={css({
-                border: '1px solid',
-                borderColor: 'border',
-                borderRadius: '8px',
-                bg: 'background',
-                overflow: 'hidden',
-            })}
-        >
-            <div
-                onClick={() => setExpanded((p) => !p)}
-                className={css({
-                    display: 'flex',
-                    alignItems: 'center',
-                    gap: '6px',
-                    padding: '8px 10px',
-                    cursor: 'pointer',
-                    _hover: { bg: 'surfaceHover' },
-                })}
-            >
-                <Chevron open={expanded} />
-                <span className={css({ flex: 1, fontSize: '13px', fontWeight: '600', color: 'text' })}>
-                    {component.type}
-                </span>
-                <button
-                    type="button"
-                    onClick={(e) => {
-                        e.stopPropagation();
-                        onDelete();
-                    }}
-                    title="このコンポーネントを削除"
-                    className={css({
-                        width: '22px',
-                        height: '22px',
-                        bg: 'transparent',
-                        color: 'errorText',
-                        border: '1px solid',
-                        borderColor: 'border',
-                        borderRadius: '4px',
-                        fontSize: '12px',
-                        cursor: 'pointer',
-                        _hover: { borderColor: 'errorLight' },
-                    })}
-                >
-                    ×
-                </button>
-            </div>
-            {expanded && (
-                <div
-                    className={css({
-                        padding: '10px',
-                        borderTop: '1px solid',
-                        borderColor: 'border',
-                        bg: 'surface',
-                    })}
-                >
-                    <ComponentDataEditor
-                        component={component}
-                        componentIndex={componentIndex}
-                        dataFields={dataFields}
-                        onChange={onChange}
-                    />
-                </div>
-            )}
-        </div>
-    );
-}
-
-// ============================================
-// Component data エディタ (フォーム + JSON タブ)
-// ============================================
-
-function ComponentDataEditor({
-    component,
-    componentIndex,
-    dataFields,
-    onChange,
-}: {
-    component: EntityComponentDef;
-    componentIndex: number;
-    dataFields?: DataFields;
-    onChange: (updater: (prev: InitialEntity) => InitialEntity) => void;
-}) {
-    const [dataTab, setDataTab] = useState<'form' | 'json'>('form');
-    const [jsonText, setJsonText] = useState(() => JSON.stringify(component.data ?? {}, null, 2));
-    const [jsonError, setJsonError] = useState('');
-
-    useEffect(() => {
-        setJsonText(JSON.stringify(component.data ?? {}, null, 2));
-        setJsonError('');
-    }, [component]);
-
-    const data = (component.data as Record<string, unknown> | undefined) ?? {};
-    const setData = (next: Record<string, unknown>) => {
-        onChange((prev) => ({
-            ...prev,
-            components: prev.components.map((c, i) => (i === componentIndex ? { ...c, data: next } : c)),
-        }));
-        setJsonText(JSON.stringify(next, null, 2));
-        setJsonError('');
-    };
-
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '2' })}>
-            <div className={css({ display: 'flex', gap: '4px', alignSelf: 'flex-end' })}>
-                <MiniTab active={dataTab === 'form'} onClick={() => setDataTab('form')} label="フォーム" />
-                <MiniTab active={dataTab === 'json'} onClick={() => setDataTab('json')} label="JSON" />
-            </div>
-            {dataTab === 'form' ? (
-                <DataFormFields data={data} dataFields={dataFields} onChange={setData} />
-            ) : (
-                <DataJsonField
-                    text={jsonText}
-                    error={jsonError}
-                    onChange={(text) => {
-                        setJsonText(text);
-                        try {
-                            const parsed = JSON.parse(text || '{}') as Record<string, unknown>;
-                            onChange((prev) => ({
-                                ...prev,
-                                components: prev.components.map((c, i) =>
-                                    i === componentIndex ? { ...c, data: parsed } : c,
-                                ),
-                            }));
-                            setJsonError('');
-                        } catch (err) {
-                            setJsonError(err instanceof Error ? err.message : 'JSON parse error');
-                        }
-                    }}
-                />
-            )}
-        </div>
-    );
-}
-
-// ============================================
-// Component 追加ピッカー (重複可)
-// ============================================
-
-function ComponentPicker({
-    availableKinds,
-    onAdd,
-}: {
-    availableKinds: AvailableEntityKind[];
-    onAdd: (type: string) => void;
-}) {
-    const [type, setType] = useState('');
-    const [dragOver, setDragOver] = useState(false);
-    const candidates = useMemo(() => availableKinds.map((k) => k.kind).sort(), [availableKinds]);
-
-    const handleAdd = () => {
-        if (!type) return;
-        onAdd(type);
-        setType('');
-    };
-
-    return (
-        <div
-            onDragOver={(e) => {
-                if (e.dataTransfer.types.includes(COMPONENT_DRAG_MIME)) {
-                    e.preventDefault();
-                    e.dataTransfer.dropEffect = 'copy';
-                    setDragOver(true);
-                }
-            }}
-            onDragLeave={() => setDragOver(false)}
-            onDrop={(e) => {
-                const ctype = e.dataTransfer.getData(COMPONENT_DRAG_MIME);
-                setDragOver(false);
-                if (!ctype) return;
-                e.preventDefault();
-                onAdd(ctype);
-            }}
-            className={css({
-                display: 'grid',
-                gridTemplateColumns: '1fr auto',
-                gap: '6px',
-                pt: '2',
-                borderTop: '1px dashed',
-                borderColor: 'border',
-                outline: dragOver ? '2px dashed' : 'none',
-                outlineColor: 'primary',
-                outlineOffset: '2px',
-                borderRadius: '4px',
-            })}
-        >
-            <select value={type} onChange={(e) => setType(e.target.value)} className={inputStyle}>
-                <option value="">+ コンポーネントを選択...</option>
-                {candidates.map((k) => (
-                    <option key={k} value={k}>
-                        {k}
-                    </option>
-                ))}
-            </select>
-            <button
-                type="button"
-                onClick={handleAdd}
-                disabled={!type}
-                className={css({
-                    padding: '6px 12px',
-                    bg: 'primary',
-                    color: 'textOnPrimary',
-                    border: 'none',
-                    borderRadius: '6px',
-                    fontSize: '12px',
-                    fontWeight: '600',
-                    cursor: 'pointer',
-                    _disabled: { opacity: 0.4, cursor: 'not-allowed' },
-                    _hover: { opacity: 0.9 },
-                })}
-            >
-                追加
-            </button>
-        </div>
-    );
-}
-
-// ============================================
-// Tags
-// ============================================
-
-const TAG_PATTERN = /^[a-z0-9_-]+$/;
-
-function TagsEditor({ tags, onChange }: { tags: string[]; onChange: (next: string[]) => void }) {
-    const [draft, setDraft] = useState('');
-    const [error, setError] = useState('');
-
-    const addTag = () => {
-        const next = draft.trim().toLowerCase();
-        if (!next) return;
-        if (!TAG_PATTERN.test(next)) {
-            setError('小文字英数 + - _ のみ');
-            return;
-        }
-        if (tags.includes(next)) {
-            setError('既に追加済み');
-            return;
-        }
-        onChange([...tags, next]);
-        setDraft('');
-        setError('');
-    };
-
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '2' })}>
-            {tags.length > 0 && (
-                <div className={css({ display: 'flex', flexWrap: 'wrap', gap: '4px' })}>
-                    {tags.map((t) => (
-                        <span
-                            key={t}
-                            className={css({
-                                display: 'inline-flex',
-                                alignItems: 'center',
-                                gap: '4px',
-                                padding: '3px 6px 3px 8px',
-                                bg: 'primarySubtle',
-                                color: 'primary',
-                                borderRadius: '999px',
-                                fontSize: '11px',
-                                fontWeight: '600',
-                            })}
-                        >
-                            {t}
-                            <button
-                                type="button"
-                                onClick={() => onChange(tags.filter((x) => x !== t))}
-                                aria-label={`${t} を削除`}
-                                className={css({
-                                    width: '16px',
-                                    height: '16px',
-                                    bg: 'transparent',
-                                    border: 'none',
-                                    color: 'primary',
-                                    cursor: 'pointer',
-                                    fontSize: '12px',
-                                    lineHeight: '1',
-                                    _hover: { opacity: 0.7 },
-                                })}
-                            >
-                                ×
-                            </button>
-                        </span>
-                    ))}
-                </div>
-            )}
-            <div className={css({ display: 'grid', gridTemplateColumns: '1fr auto', gap: '6px' })}>
-                <input
-                    type="text"
-                    value={draft}
-                    onChange={(e) => {
-                        setDraft(e.target.value);
-                        setError('');
-                    }}
-                    onKeyDown={(e) => {
-                        if (e.key === 'Enter') {
-                            e.preventDefault();
-                            addTag();
-                        }
-                    }}
-                    placeholder="tag を追加 (例: ui, background)"
-                    className={inputStyle}
-                />
-                <button
-                    type="button"
-                    onClick={addTag}
-                    disabled={!draft.trim()}
-                    className={css({
-                        padding: '6px 12px',
-                        bg: 'primary',
-                        color: 'textOnPrimary',
-                        border: 'none',
-                        borderRadius: '6px',
-                        fontSize: '12px',
-                        fontWeight: '600',
-                        cursor: 'pointer',
-                        _disabled: { opacity: 0.4, cursor: 'not-allowed' },
-                        _hover: { opacity: 0.9 },
-                    })}
-                >
-                    + 追加
-                </button>
-            </div>
-            {error && <span className={css({ fontSize: '11px', color: 'errorText' })}>{error}</span>}
-        </div>
-    );
-}
-
-// ============================================
-// Entity ヘッダー: 名前リネーム + 削除
-// ============================================
-
-function EntityHeader({
-    entity,
-    onDelete,
-    onRename,
-}: {
-    entity: InitialEntity;
-    onDelete: () => void;
-    onRename: (id: string) => void;
-}) {
-    const [editing, setEditing] = useState(false);
-    const [draft, setDraft] = useState(entity.id);
-    const inputRef = useRef<HTMLInputElement>(null);
-
-    useEffect(() => {
-        if (!editing) setDraft(entity.id);
-    }, [entity.id, editing]);
-
-    useEffect(() => {
-        if (editing) inputRef.current?.focus();
-    }, [editing]);
-
-    const commit = () => {
-        const next = draft.trim();
-        if (next && next !== entity.id) onRename(next);
-        setEditing(false);
-    };
-
-    return (
-        <div className={css({ display: 'flex', justifyContent: 'space-between', alignItems: 'center', gap: '2' })}>
-            <div className={css({ flex: 1, minW: 0 })}>
-                {editing ? (
-                    <input
-                        ref={inputRef}
-                        type="text"
-                        value={draft}
-                        onChange={(e) => setDraft(e.target.value)}
-                        onBlur={commit}
-                        onKeyDown={(e) => {
-                            if (e.key === 'Enter') commit();
-                            else if (e.key === 'Escape') {
-                                setDraft(entity.id);
-                                setEditing(false);
-                            }
-                        }}
-                        className={inputStyle}
-                    />
-                ) : (
-                    <button
-                        type="button"
-                        onClick={() => setEditing(true)}
-                        title="クリックして ID を編集"
-                        className={css({
-                            fontSize: '14px',
-                            fontWeight: '700',
-                            color: 'text',
-                            bg: 'transparent',
-                            border: 'none',
-                            cursor: 'pointer',
-                            padding: '4px 0',
-                            textAlign: 'left',
-                            width: '100%',
-                            overflow: 'hidden',
-                            textOverflow: 'ellipsis',
-                            whiteSpace: 'nowrap',
-                            _hover: { color: 'primary' },
-                        })}
-                    >
-                        {entity.id}
-                    </button>
-                )}
-            </div>
-            <button
-                type="button"
-                onClick={onDelete}
-                className={css({
-                    padding: '6px 12px',
-                    bg: 'errorBg',
-                    color: 'errorText',
-                    border: '1px solid',
-                    borderColor: 'errorLight',
-                    borderRadius: '8px',
-                    fontSize: '12px',
-                    fontWeight: '600',
-                    cursor: 'pointer',
-                    _hover: { opacity: 0.9 },
-                })}
-            >
-                削除
-            </button>
-        </div>
-    );
-}
-
-// ============================================
-// Data フォーム編集
-// ============================================
-
-function DataFormFields({
-    data,
-    dataFields,
-    onChange,
-}: {
-    data: Record<string, unknown>;
-    dataFields?: DataFields;
-    onChange: (next: Record<string, unknown>) => void;
-}) {
-    const setField = (key: string, value: unknown) => onChange({ ...data, [key]: value });
-    const deleteField = (key: string) => {
-        const next = { ...data };
-        delete next[key];
-        onChange(next);
-    };
-
-    const declaredKeys = useMemo(() => Object.keys(dataFields ?? {}), [dataFields]);
-    const customKeys = useMemo(() => {
-        const declared = new Set(declaredKeys);
-        return Object.keys(data).filter((k) => !declared.has(k));
-    }, [data, declaredKeys]);
-
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '3' })}>
-            {declaredKeys.length > 0 && (
-                <div className={css({ display: 'flex', flexDirection: 'column', gap: '3' })}>
-                    {declaredKeys.map((key) => {
-                        const spec = dataFields?.[key];
-                        if (!spec) return null;
-                        const value = key in data ? data[key] : (spec.default ?? defaultForType(spec));
-                        return (
-                            <DeclaredFieldRow
-                                key={key}
-                                fieldKey={key}
-                                spec={spec}
-                                value={value}
-                                onChange={(v) => setField(key, v)}
-                                onReset={() => setField(key, spec.default ?? defaultForType(spec))}
-                            />
-                        );
-                    })}
-                </div>
-            )}
-
-            {customKeys.length > 0 && (
-                <div
-                    className={css({
-                        display: 'flex',
-                        flexDirection: 'column',
-                        gap: '2',
-                        pt: '3',
-                        borderTop: declaredKeys.length > 0 ? '1px dashed' : 'none',
-                        borderColor: 'border',
-                    })}
-                >
-                    {declaredKeys.length > 0 && (
-                        <span
-                            className={css({
-                                fontSize: '11px',
-                                color: 'textSubtle',
-                                fontWeight: '600',
-                                textTransform: 'uppercase',
-                                letterSpacing: '0.04em',
-                            })}
-                        >
-                            カスタム
-                        </span>
-                    )}
-                    {customKeys.map((key) => (
-                        <CustomFieldRow
-                            key={key}
-                            fieldKey={key}
-                            value={data[key]}
-                            onChange={(v) => setField(key, v)}
-                            onDelete={() => deleteField(key)}
-                        />
-                    ))}
-                </div>
-            )}
-
-            {declaredKeys.length === 0 && customKeys.length === 0 && (
-                <div className={css({ fontSize: '12px', color: 'textSubtle' })}>
-                    {dataFields === undefined
-                        ? 'プラグインが data フィールドを宣言していません。下から自由にキーを追加できます（プラグインが読まない可能性があります）。'
-                        : 'このコンポーネントはデータを必要としません。'}
-                </div>
-            )}
-
-            {dataFields === undefined && (
-                <AddFieldRow existing={new Set(Object.keys(data))} onAdd={(k, v) => onChange({ ...data, [k]: v })} />
-            )}
-        </div>
-    );
-}
-
-function DeclaredFieldRow({
-    fieldKey,
-    spec,
-    value,
-    onChange,
-    onReset,
-}: {
-    fieldKey: string;
-    spec: DataFieldSpec;
-    value: unknown;
-    onChange: (v: unknown) => void;
-    onReset: () => void;
-}) {
-    const label = spec.label ?? fieldKey;
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '4px' })}>
-            <div className={css({ display: 'flex', alignItems: 'center', justifyContent: 'space-between' })}>
-                <span className={css({ fontSize: '12px', fontWeight: '600', color: 'text' })}>
-                    {label}
-                    <span className={css({ ml: '6px', fontSize: '10px', color: 'textSubtle', fontWeight: '500' })}>
-                        ({fieldKey} · {spec.type})
-                    </span>
-                </span>
-                <button
-                    type="button"
-                    onClick={onReset}
-                    title="既定値に戻す"
-                    className={css({
-                        fontSize: '10px',
-                        color: 'textSubtle',
-                        bg: 'transparent',
-                        border: 'none',
-                        cursor: 'pointer',
-                        _hover: { color: 'primary' },
-                    })}
-                >
-                    ↻
-                </button>
-            </div>
-            <DeclaredInput spec={spec} value={value} onChange={onChange} />
-            {spec.help && <span className={css({ fontSize: '10px', color: 'textSubtle' })}>{spec.help}</span>}
-        </div>
-    );
-}
-
-function DeclaredInput({
-    spec,
-    value,
-    onChange,
-}: {
-    spec: DataFieldSpec;
-    value: unknown;
-    onChange: (v: unknown) => void;
-}) {
-    if (spec.type === 'string') {
-        const s = String(value ?? '');
-        if (spec.multiline) {
-            return (
-                <textarea
-                    value={s}
-                    onChange={(e) => onChange(e.target.value)}
-                    placeholder={spec.placeholder}
-                    rows={3}
-                    className={textareaStyle}
-                />
-            );
-        }
-        return (
-            <input
-                type="text"
-                value={s}
-                onChange={(e) => onChange(e.target.value)}
-                placeholder={spec.placeholder}
-                className={inputStyle}
-            />
-        );
-    }
-    if (spec.type === 'url') {
-        return (
-            <input
-                type="url"
-                value={String(value ?? '')}
-                onChange={(e) => onChange(e.target.value)}
-                placeholder={spec.placeholder ?? 'https://...'}
-                className={inputStyle}
-            />
-        );
-    }
-    if (spec.type === 'number') {
-        return (
-            <NumberInput
-                value={typeof value === 'number' ? value : 0}
-                min={spec.min}
-                max={spec.max}
-                step={spec.step ?? 1}
-                onChange={onChange}
-                className={inputStyle}
-            />
-        );
-    }
-    if (spec.type === 'boolean') {
-        const v = Boolean(value);
-        return (
-            <label className={css({ display: 'inline-flex', alignItems: 'center', gap: '8px', fontSize: '12px' })}>
-                <input type="checkbox" checked={v} onChange={(e) => onChange(e.target.checked)} />
-                <span className={css({ color: 'textMuted' })}>{v ? 'true' : 'false'}</span>
-            </label>
-        );
-    }
-    if (spec.type === 'color') {
-        const v = typeof value === 'string' ? value : '#000000';
-        return (
-            <div className={css({ display: 'flex', gap: '6px', alignItems: 'center' })}>
-                <input
-                    type="color"
-                    value={v}
-                    onChange={(e) => onChange(e.target.value)}
-                    className={css({
-                        width: '40px',
-                        height: '32px',
-                        borderRadius: '6px',
-                        border: '1px solid',
-                        borderColor: 'border',
-                        cursor: 'pointer',
-                    })}
-                />
-                <input type="text" value={v} onChange={(e) => onChange(e.target.value)} className={inputStyle} />
-            </div>
-        );
-    }
-    if (spec.type === 'enum') {
-        return (
-            <select value={String(value ?? '')} onChange={(e) => onChange(e.target.value)} className={inputStyle}>
-                {spec.options.map((o) => (
-                    <option key={o} value={o}>
-                        {o}
-                    </option>
-                ))}
-            </select>
-        );
-    }
-    return <MiniJsonEditor value={value} onChange={onChange} />;
-}
-
-function CustomFieldRow({
-    fieldKey,
-    value,
-    onChange,
-    onDelete,
-}: {
-    fieldKey: string;
-    value: unknown;
-    onChange: (v: unknown) => void;
-    onDelete: () => void;
-}) {
-    const type = detectType(value);
-    return (
-        <div
-            className={css({
-                display: 'grid',
-                gridTemplateColumns: '110px 1fr auto',
-                gap: '6px',
-                alignItems: 'start',
-            })}
-        >
-            <span
-                className={css({
-                    fontSize: '12px',
-                    color: 'text',
-                    fontWeight: '600',
-                    pt: '6px',
-                    overflow: 'hidden',
-                    textOverflow: 'ellipsis',
-                    whiteSpace: 'nowrap',
-                })}
-                title={fieldKey}
-            >
-                {fieldKey}
-            </span>
-            {type === 'boolean' ? (
-                <label className={css({ display: 'inline-flex', alignItems: 'center', gap: '6px', fontSize: '12px' })}>
-                    <input type="checkbox" checked={Boolean(value)} onChange={(e) => onChange(e.target.checked)} />
-                    <span className={css({ color: 'textMuted' })}>{value ? 'true' : 'false'}</span>
-                </label>
-            ) : type === 'number' ? (
-                <NumberInput value={Number(value)} onChange={onChange} className={inputStyle} />
-            ) : type === 'string' ? (
-                <input
-                    type="text"
-                    value={String(value)}
-                    onChange={(e) => onChange(e.target.value)}
-                    className={inputStyle}
-                />
-            ) : (
-                <MiniJsonEditor value={value} onChange={onChange} />
-            )}
-            <button
-                type="button"
-                onClick={onDelete}
-                title="このフィールドを削除"
-                className={css({
-                    padding: '4px 8px',
-                    bg: 'transparent',
-                    color: 'errorText',
-                    border: '1px solid',
-                    borderColor: 'border',
-                    borderRadius: '6px',
-                    fontSize: '11px',
-                    cursor: 'pointer',
-                    _hover: { borderColor: 'errorLight' },
-                })}
-            >
-                ×
-            </button>
-        </div>
-    );
-}
-
-function MiniJsonEditor({ value, onChange }: { value: unknown; onChange: (v: unknown) => void }) {
-    const [text, setText] = useState(() => JSON.stringify(value, null, 2));
-    const [err, setErr] = useState('');
-    useEffect(() => {
-        setText(JSON.stringify(value, null, 2));
-        setErr('');
-    }, [value]);
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '2px' })}>
-            <textarea
-                value={text}
-                onChange={(e) => {
-                    setText(e.target.value);
-                    try {
-                        const parsed = JSON.parse(e.target.value || 'null') as unknown;
-                        onChange(parsed);
-                        setErr('');
-                    } catch (er) {
-                        setErr(er instanceof Error ? er.message : 'JSON parse error');
-                    }
-                }}
-                rows={3}
-                spellCheck={false}
-                className={textareaStyle}
-            />
-            {err && <span className={css({ fontSize: '10px', color: 'errorText' })}>{err}</span>}
-        </div>
-    );
-}
-
-function AddFieldRow({ existing, onAdd }: { existing: Set<string>; onAdd: (name: string, value: unknown) => void }) {
-    const [name, setName] = useState('');
-    const [type, setType] = useState<'string' | 'number' | 'boolean' | 'json'>('string');
-
-    const handleAdd = () => {
-        const trimmed = name.trim();
-        if (!trimmed || existing.has(trimmed)) return;
-        const initial: unknown = type === 'string' ? '' : type === 'number' ? 0 : type === 'boolean' ? false : {};
-        onAdd(trimmed, initial);
-        setName('');
-        setType('string');
-    };
-
-    return (
-        <div
-            className={css({
-                display: 'grid',
-                gridTemplateColumns: '110px 1fr auto auto',
-                gap: '6px',
-                pt: '3',
-                borderTop: '1px dashed',
-                borderColor: 'border',
-            })}
-        >
-            <input
-                type="text"
-                value={name}
-                onChange={(e) => setName(e.target.value)}
-                onKeyDown={(e) => {
-                    if (e.key === 'Enter') {
-                        e.preventDefault();
-                        handleAdd();
-                    }
-                }}
-                placeholder="フィールド名"
-                className={inputStyle}
-            />
-            <select value={type} onChange={(e) => setType(e.target.value as typeof type)} className={inputStyle}>
-                <option value="string">文字列</option>
-                <option value="number">数値</option>
-                <option value="boolean">真偽値</option>
-                <option value="json">JSON</option>
-            </select>
-            <button
-                type="button"
-                onClick={handleAdd}
-                disabled={!name.trim() || existing.has(name.trim())}
-                className={css({
-                    padding: '6px 12px',
-                    bg: 'primary',
-                    color: 'textOnPrimary',
-                    border: 'none',
-                    borderRadius: '6px',
-                    fontSize: '11px',
-                    fontWeight: '600',
-                    cursor: 'pointer',
-                    _disabled: { opacity: 0.4, cursor: 'not-allowed' },
-                    _hover: { opacity: 0.9 },
-                })}
-            >
-                + 追加
-            </button>
-            <span />
-        </div>
-    );
-}
-
-function DataJsonField({ text, error, onChange }: { text: string; error: string; onChange: (text: string) => void }) {
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '2px' })}>
-            <textarea
-                value={text}
-                onChange={(e) => onChange(e.target.value)}
-                rows={8}
-                spellCheck={false}
-                className={css({
-                    width: '100%',
-                    padding: '8px 10px',
-                    borderRadius: '8px',
-                    border: '1.5px solid',
-                    borderColor: 'border',
-                    bg: 'background',
-                    color: 'text',
-                    fontFamily: 'mono',
-                    fontSize: '12px',
-                    outline: 'none',
-                    resize: 'vertical',
-                    _focus: { borderColor: 'primary' },
-                })}
-            />
-            {error && <span className={css({ fontSize: '11px', color: 'errorText' })}>{error}</span>}
-        </div>
-    );
-}
-
-// ============================================
-// 共通: NumField / Section / MiniTab / Chevron
-// ============================================
-
-/**
- * 数値入力の共通基盤。
- *  - **NaN を 0 に丸めない** (途中 "-" / "." / 空 で値が 0 にされない)
- *  - 入力中はローカルテキスト state を保持し、有効な数値のときだけ親へ commit
- *  - 親 (= 外部) からの value 変化はテキストへ同期 (例: ステージで drag したとき)
- *  - blur 時は最後に commit した値へテキストを揃える (空のまま離れない)
- */
-function NumberInput({
-    value,
-    onChange,
-    min,
-    max,
-    step,
-    className,
-}: {
-    value: number;
-    onChange: (v: number) => void;
-    min?: number;
-    max?: number;
-    step?: number;
-    className?: string;
-}) {
-    const [text, setText] = useState(String(value));
-    const lastEmittedRef = useRef(value);
-    useEffect(() => {
-        if (value !== lastEmittedRef.current) {
-            setText(String(value));
-            lastEmittedRef.current = value;
-        }
-    }, [value]);
-    const commit = (raw: string) => {
-        const n = Number.parseFloat(raw);
-        if (!Number.isFinite(n)) return; // 空 / "-" / "." 等は親に伝えない (タイプ継続を許す)
-        let v = n;
-        if (min !== undefined) v = Math.max(v, min);
-        if (max !== undefined) v = Math.min(v, max);
-        if (v !== lastEmittedRef.current) {
-            lastEmittedRef.current = v;
-            onChange(v);
-        }
-    };
-    return (
-        <input
-            type="number"
-            value={text}
-            min={min}
-            max={max}
-            step={step}
-            onChange={(e) => {
-                setText(e.target.value);
-                commit(e.target.value);
-            }}
-            onBlur={() => {
-                // 空のまま離れたら最後に commit した値を表示
-                if (text === '' || !Number.isFinite(Number.parseFloat(text))) {
-                    setText(String(lastEmittedRef.current));
-                }
-            }}
-            className={className}
-        />
-    );
-}
-
-function NumField({
-    label,
-    value,
-    max,
-    onChange,
-}: {
-    label: string;
-    value: number;
-    max?: number;
-    onChange: (v: number) => void;
-}) {
-    return (
-        // biome-ignore lint/a11y/noLabelWithoutControl: NumberInput が input を内包している (label/input 関係は成立)
-        <label className={css({ display: 'flex', flexDirection: 'column', gap: '1' })}>
-            <span className={css({ fontSize: '11px', color: 'textMuted' })}>{label}</span>
-            <NumberInput value={value} max={max} onChange={onChange} className={inputStyle} />
-        </label>
-    );
-}
-
-function Section({ label, children }: { label: string; children: React.ReactNode }) {
-    return (
-        <div className={css({ display: 'flex', flexDirection: 'column', gap: '2' })}>
-            <div className={css({ fontSize: '12px', fontWeight: '600', color: 'text' })}>{label}</div>
-            {children}
-        </div>
-    );
-}
-
-function MiniTab({ active, onClick, label }: { active: boolean; onClick: () => void; label: string }) {
-    return (
-        <button
-            type="button"
-            onClick={onClick}
-            className={css({
-                padding: '3px 8px',
-                bg: active ? 'primary' : 'background',
-                color: active ? 'textOnPrimary' : 'textMuted',
-                border: '1px solid',
-                borderColor: active ? 'primary' : 'border',
-                borderRadius: '6px',
-                fontSize: '11px',
-                fontWeight: '600',
-                cursor: 'pointer',
-                _hover: { opacity: 0.9 },
-            })}
-        >
-            {label}
-        </button>
-    );
-}
-
-function Chevron({ open }: { open: boolean }) {
-    return (
-        <svg
-            width="10"
-            height="10"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            strokeWidth="2.5"
-            style={{ transform: open ? 'rotate(90deg)' : 'rotate(0deg)', transition: 'transform 0.15s' }}
-            aria-hidden="true"
-        >
-            <path d="M9 18l6-6-6-6" />
-        </svg>
-    );
-}
-
-function detectType(v: unknown): 'string' | 'number' | 'boolean' | 'json' {
-    if (typeof v === 'string') return 'string';
-    if (typeof v === 'number') return 'number';
-    if (typeof v === 'boolean') return 'boolean';
-    return 'json';
-}
-
-function defaultForType(spec: DataFieldSpec): unknown {
-    switch (spec.type) {
-        case 'string':
-        case 'url':
-            return '';
-        case 'number':
-            return 0;
-        case 'boolean':
-            return false;
-        case 'color':
-            return '#000000';
-        case 'enum':
-            return spec.options[0] ?? '';
-        case 'json':
-            return null;
-    }
-}
-
-const inputStyle = css({
-    width: '100%',
-    padding: '6px 8px',
-    borderRadius: '6px',
-    border: '1px solid',
-    borderColor: 'border',
-    bg: 'background',
-    color: 'text',
-    fontSize: '12px',
-    outline: 'none',
-    _focus: { borderColor: 'primary' },
-});
-
-const textareaStyle = css({
-    width: '100%',
-    padding: '6px 8px',
-    borderRadius: '6px',
-    border: '1px solid',
-    borderColor: 'border',
-    bg: 'background',
-    color: 'text',
-    fontFamily: 'mono',
-    fontSize: '11px',
-    outline: 'none',
-    resize: 'vertical',
-    _focus: { borderColor: 'primary' },
-});
